drm/vc4: Advertise specific column heights for BROADCOM_SAND modifier
The vc4 HVS can handle almost any column height in the SAND format,
and this was signalled via the vendor bits in DRM_FORMAT_MOD_BROADCOM_SAND128.

DRM is now requiring that all potential modifier values are reported,
not just the canonical version of them.
It is not feasible to report all the potential column heights.
After review, add the aligned column heights that follow the optimal
SDRAM page allocation algorithm for a spread of image heights, and
explicity advertise those.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>